remove tons of -pedantic warnings
git-svn-id: https://insight.labri.fr/svn/trunk@202 00316523-1ed1-4fdc-9313-76483a3f41dc

+++ src/kernel/microcode/MicrocodeNode.hh
@@ -193,7 +193,7 @@ public:
 
   struct lt_arrow {
     bool operator()(StmtArrow *n1, StmtArrow *n2) const {
-      return ((long long int) n1) < ((long long int) n2) ;
+      return ((long) n1) < ((long) n2) ;
     }
   };
 };
